refactor(reporting): keep batch boundaries modular
